cgetctest: Show that cgetc() from conio does not care about color RAM cursor
In a comment[1] on cc65 issue #946, @greg-king5 describes the problem
well but says that cgetc() also needs the constructor that would fix
the color RAM cursor problem. I am not clear on why this is, since
cgetc() alone doesn't produce any output. This program is an attempt
to demonstrate that I think it's ok not to initialize the color RAM
cursor position.

[1]: https://github.com/cc65/cc65/issues/946#issuecomment-538502820